+++ zetaclient/chains/base/confirmation_test.go
@@ -1,10 +1,18 @@
 package base_test
 
 import (
+	"context"
+	"errors"
 	"testing"
 
+	sdkmath "cosmossdk.io/math"
+	ethcommon "github.com/ethereum/go-ethereum/common"
+	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 	"github.com/zeta-chain/node/pkg/chains"
+	"github.com/zeta-chain/node/pkg/coin"
+	crosschaintypes "github.com/zeta-chain/node/x/crosschain/types"
+	fungibletypes "github.com/zeta-chain/node/x/fungible/types"
 	observertypes "github.com/zeta-chain/node/x/observer/types"
 )
 
@@ -27,7 +35,7 @@ func Test_GetScanRangeInboundSafe(t *testing.T) {
 			confParams: observertypes.ConfirmationParams{
 				SafeInboundCount: 10,
 			},
-			expectedBlockRange: [2]uint64{0, 0}, // [0, 0)
+			expectedBlockRange: [2]uint64{91, 91}, // [91, 91), nothing to scan
 		},
 		{
 			name:        "1 unscanned blocks",
@@ -93,7 +101,7 @@ func Test_GetScanRangeInboundFast(t *testing.T) {
 				SafeInboundCount: 10,
 				FastInboundCount: 10,
 			},
-			expectedBlockRange: [2]uint64{0, 0}, // [0, 0)
+			expectedBlockRange: [2]uint64{91, 91}, // [91, 91), nothing to scan
 		},
 		{
 			name:        "1 unscanned blocks",
@@ -295,6 +303,112 @@ func Test_IsBlockConfirmedForOutboundFast(t *testing.T) {
 
 			isConfirmed := ob.IsBlockConfirmedForOutboundFast(tt.blockNumber)
 			require.Equal(t, tt.expected, isConfirmed)
+		})
+	}
+}
+
+func Test_IsInboundEligibleForFastConfirmation(t *testing.T) {
+	chain := chains.Ethereum
+	liquidityCap := sdkmath.NewUint(100_000)
+	fastAmountCap := chains.CalcInboundFastConfirmationAmountCap(chain.ChainId, liquidityCap)
+	confParamsEnabled := observertypes.ConfirmationParams{
+		SafeInboundCount: 2,
+		FastInboundCount: 1,
+	}
+
+	tests := []struct {
+		name                string
+		confParams          observertypes.ConfirmationParams
+		msg                 *crosschaintypes.MsgVoteInbound
+		failForeignCoinsRPC bool
+		eligible            bool
+		errMsg              string
+	}{
+		{
+			name:       "eligible for fast confirmation",
+			confParams: confParamsEnabled,
+			msg: &crosschaintypes.MsgVoteInbound{
+				SenderChainId:           chain.ChainId,
+				Amount:                  sdkmath.NewUint(fastAmountCap.Uint64()),
+				CoinType:                coin.CoinType_Gas,
+				Asset:                   "",
+				ProtocolContractVersion: crosschaintypes.ProtocolContractVersion_V2,
+			},
+			eligible: true,
+		},
+		{
+			name: "not eligible if fast confirmation is disabled",
+			confParams: observertypes.ConfirmationParams{
+				SafeInboundCount: 2,
+				FastInboundCount: 2, // equal to safe confirmation, effectively disabled
+			},
+			msg: &crosschaintypes.MsgVoteInbound{
+				SenderChainId: chains.SolanaMainnet.ChainId, // not set for Solana
+			},
+			eligible: false,
+		},
+		{
+			name:       "not eligible if protocol contract version V1 is used",
+			confParams: confParamsEnabled,
+			msg: &crosschaintypes.MsgVoteInbound{
+				SenderChainId:           chain.ChainId,
+				ProtocolContractVersion: crosschaintypes.ProtocolContractVersion_V1, // not eligible for V1
+			},
+			eligible: false,
+		},
+		{
+			name:       "return error if foreign coins query RPC fails",
+			confParams: confParamsEnabled,
+			msg: &crosschaintypes.MsgVoteInbound{
+				SenderChainId:           chain.ChainId,
+				CoinType:                coin.CoinType_Gas,
+				Asset:                   "",
+				ProtocolContractVersion: crosschaintypes.ProtocolContractVersion_V2,
+			},
+			failForeignCoinsRPC: true,
+			eligible:            false,
+			errMsg:              "unable to get foreign coins",
+		},
+		{
+			name:       "not eligible if amount exceeds fast amount cap",
+			confParams: confParamsEnabled,
+			msg: &crosschaintypes.MsgVoteInbound{
+				SenderChainId:           chain.ChainId,
+				Amount:                  sdkmath.NewUint(fastAmountCap.Uint64() + 1), // +1 to exceed
+				CoinType:                coin.CoinType_Gas,
+				Asset:                   "",
+				ProtocolContractVersion: crosschaintypes.ProtocolContractVersion_V2,
+			},
+			eligible: false,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			// ARRANGE
+			ob := newTestSuite(t, chain, withConfirmationParams(tt.confParams))
+
+			// mock up the foreign coins RPC
+			assetAddress := ethcommon.HexToAddress(tt.msg.Asset)
+			if tt.failForeignCoinsRPC {
+				ob.zetacore.On("GetForeignCoinsFromAsset", mock.Anything, chain.ChainId, assetAddress).
+					Maybe().
+					Return(fungibletypes.ForeignCoins{}, errors.New("rpc failed"))
+			} else {
+				ob.zetacore.On("GetForeignCoinsFromAsset", mock.Anything, chain.ChainId, assetAddress).Maybe().Return(fungibletypes.ForeignCoins{LiquidityCap: liquidityCap}, nil)
+			}
+
+			// ACT
+			ctx := context.Background()
+			eligible, err := ob.IsInboundEligibleForFastConfirmation(ctx, tt.msg)
+
+			// ASSERT
+			require.Equal(t, tt.eligible, eligible)
+			if tt.errMsg != "" {
+				require.Contains(t, err.Error(), tt.errMsg)
+				return
+			}
+			require.NoError(t, err)
 		})
 	}
 }